requirements.txt
# Notion Backup Automation - Python Dependencies

# Notion SDK - Official Python client for Notion API
notion-client==2.2.1

# Environment variable management
python-dotenv==1.0.0

# Optional: For Google Drive integration (uncomment if needed)
# google-auth==2.23.0
# google-auth-oauthlib==1.1.0
# google-auth-httplib2==0.1.1
# google-api-python-client==2.100.0

# Optional: For enhanced notifications (uncomment if needed)
# requests==2.31.0  # For Slack/Discord webhooks
# sendgrid==6.10.0  # For email notifications
